requirements.txt
autopep8==1.7.0
Pillow==9.2.0
pycodestyle==2.9.1
toml==0.10.2
